pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-toolkit"
version = "0.1.0"
description = "MedRAG starter project with a shared RAG core and MedRAG-specific ingestion."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "deepeval>=2.0,<3",
  "fastapi>=0.115,<1",
  "fastembed>=0.4,<1",
  "httpx>=0.28,<0.29",
  "llama-index-core>=0.12,<0.13",
  "llama-index-embeddings-fastembed>=0.3,<0.4",
  "llama-index-llms-openai>=0.4,<0.5",
  "llama-index-readers-papers>=0.3,<0.4",
  "llama-index-vector-stores-qdrant>=0.4,<0.5",
  "llama-parse>=0.6,<0.7",
  "pydantic>=2.9,<3",
  "python-dotenv>=1.0,<2",
  "python-multipart>=0.0.9,<1",
  "qdrant-client>=1.12,<2",
  "requests>=2.32,<3",
  "streamlit>=1.40,<2",
  "uvicorn[standard]>=0.30,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
  "pytest-cov>=6,<7",
  "ruff>=0.11,<0.12",
]

[project.scripts]
rag-toolkit = "src.cli:main"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
markers = [
  "integration: integration tests that require external services or models",
]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
